+++ config.php
@@ -1,6 +1,9 @@
 <?php
 
+use humhub\commands\CronController;
 use humhub\modules\admin\widgets\AdminMenu;
+use humhub\modules\updater\Events;
+use humhub\modules\updater\modules\packageinstaller\Module;
 use yii\web\Application;
 
 /** @noinspection MissedFieldInspection */
@@ -9,9 +12,8 @@ return [
     'class' => 'humhub\modules\updater\Module',
     'namespace' => 'humhub\modules\updater',
     'events' => [
-        ['class' => AdminMenu::className(), 'event' => AdminMenu::EVENT_INIT, 'callback' => ['humhub\modules\updater\Events', 'onAdminMenuInit']],
-        ['class' => 'yii\web\Application', 'event' => Application::EVENT_BEFORE_REQUEST, 'callback' => ['humhub\modules\updater\modules\packageinstaller\Module', 'onApplicationInit']],
-        ['class' => '\humhub\commands\CronController', 'event'  => 'daily', 'callback' => ['humhub\modules\updater\Events', 'onCronRun']],
+        ['class' => AdminMenu::class, 'event' => AdminMenu::EVENT_INIT, 'callback' => [Events::class, 'onAdminMenuInit']],
+        ['class' => Application::class, 'event' => Application::EVENT_BEFORE_REQUEST, 'callback' => [Module::class, 'onApplicationInit']],
+        ['class' => CronController::class, 'event'  => CronController::EVENT_ON_DAILY_RUN, 'callback' => [Events::class, 'onCronRun']],
     ],
 ];
-?>