+++ controllers/AdminController.php
@@ -1,4 +1,5 @@
 <?php
+
 /**
  * @link https://www.humhub.org/
  * @copyright Copyright (c) 2018 HumHub GmbH & Co. KG
@@ -13,7 +14,6 @@ use Yii;
 
 class AdminController extends Controller
 {
-
     public function actionIndex()
     {
         $model = new ConfigureForm();