+++ libs/AvailableUpdate.php
@@ -8,6 +8,7 @@
 
 namespace humhub\modules\updater\libs;
 
+use humhub\widgets\bootstrap\Button;
 use Yii;
 
 /**
@@ -17,7 +18,6 @@ use Yii;
  */
 class AvailableUpdate
 {
-
     public $releaseNotes;
     public $versionFrom;
     public $fileName;
@@ -40,7 +40,7 @@ class AvailableUpdate
                 if (class_exists('\yii\httpclient\Client')) {
                     $fh = fopen($targetFile, 'w');
                     $client = new \yii\httpclient\Client([
-                        'transport' => 'yii\httpclient\CurlTransport'
+                        'transport' => 'yii\httpclient\CurlTransport',
                     ]);
                     $response = $client->createRequest()
                         ->setMethod('GET')
@@ -55,18 +55,18 @@ class AvailableUpdate
                     }
                 } else {
                     // Older Versions
-                    $http = new \Zend\Http\Client($this->downloadUrl, array(
+                    $http = new \Zend\Http\Client($this->downloadUrl, [
                         'adapter' => '\Zend\Http\Client\Adapter\Curl',
                         'curloptions' => Yii::$app->getModule('updater')->getCurlOptions(),
-                        'timeout' => 300
-                    ));
+                        'timeout' => 300,
+                    ]);
                     $http->setStream();
                     $response = $http->send();
                     copy($response->getStreamName(), $targetFile);
                 }
 
             } catch (\Exception $ex) {
-                throw new \Exception(Yii::t('UpdaterModule.libs_UpdatePackage', 'Update download failed! (%error%)', array('%error%' => $ex->getMessage())));
+                throw new \Exception(Yii::t('UpdaterModule.base', 'Update download failed! (%error%)', ['%error%' => $ex->getMessage()]));
             }
         }
 
@@ -75,4 +75,36 @@ class AvailableUpdate
         }
     }
 
+    public function getWarningMessage(): ?string
+    {
+        $warningMessages = [
+            '1.18'
+                => Yii::t('UpdaterModule.base', 'IMPORTANT NOTES ABOUT UPDATING TO VERSION 1.18:') . '<br><br>'
+                . Yii::t('UpdaterModule.base', 'All themes will be disabled, <code>.bs3.old</code> will be added to their names, and the default HumHub theme will be enabled.')
+                . (Yii::$app->getModule('theme-builder') ? '<br><br>' . Yii::t('UpdaterModule.base', 'The Theme Builder module will be uninstalled.') : '') . '<br><br>'
+                . Yii::t('UpdaterModule.base', 'Please read the {MigrationGuideLink}', ['MigrationGuideLink' => Button::asLink(Yii::t('UpdaterModule.base', 'Migration Guide'), 'https://docs.humhub.org/docs/theme/migrate')->options(['target' => '_blank'])]),
+        ];
+
+        foreach ($warningMessages as $warningVersion => $warningMessage) {
+            if (
+                version_compare($this->versionFrom, $warningVersion, '<')
+                && version_compare($this->versionTo, $warningVersion, '>=')
+            ) {
+                return $warningMessage;
+            }
+        }
+
+        return null;
+    }
+
+    public function hideSwitchDefaultThemeCheckbox(): bool
+    {
+        $targetVersionsToHide = [
+            '1.18' => true,
+        ];
+
+        return
+            Yii::$app->view->theme->name === 'HumHub'
+            || ($targetVersionsToHide[$this->versionTo] ?? false);
+    }
 }