+++ messages/cy/libs_UpdatePackage.php
@@ -1,5 +1,0 @@
-<?php
-
-return [
-    'Update download failed! (%error%)' => '',
-];