+++ messages/es/libs_UpdatePackage.php
@@ -1,4 +1,0 @@
-<?php
-return array (
-  'Update download failed! (%error%)' => 'Ha fallado la descarga de la actualización (%error%)',
-);