+++ messages/fa-IR/libs_UpdatePackage.php
@@ -1,4 +1,0 @@
-<?php
-return array (
-  'Update download failed! (%error%)' => 'مشكلي پيش آمد و دانلود بروزرسان انجام نشد....كد ارور(%error%) ',
-);