+++ messages/hu/libs_UpdatePackage.php
@@ -1,4 +1,0 @@
-<?php
-return array (
-  'Update download failed! (%error%)' => 'A frissítés letöltése sikertelen! (%error%)',
-);