+++ messages/nb-NO/libs_UpdatePackage.php
@@ -1,4 +1,0 @@
-<?php
-return array (
-  'Update download failed! (%error%)' => 'Nedlasting av oppdateringen mislyktes (%error%)',
-);