+++ messages/ru/libs_UpdatePackage.php
@@ -1,4 +1,0 @@
-<?php
-return array (
-  'Update download failed! (%error%)' => 'Ошибка загрузки при обновлении! (%error%)',
-);