+++ messages/zh-TW/libs_UpdatePackage.php
@@ -1,4 +1,0 @@
-<?php
-return array (
-  'Update download failed! (%error%)' => '下載更新失敗! (%error%)',
-);