+++ docs/CHANGELOG.md
@@ -1,13 +1,25 @@
 Changelog
 =========
 
-2.3.0 (Unreleased)
+2.4.0 (September 1, 2025)
 -------------------------
 - Fix #48: Update module resources path
+- Enh #56: Migration to Bootstrap 5 for HumHub 1.18
 
-2.2.2 (Unreleased)
+2.3.1 (August 6, 2025)
+----------------------
+- Enh #57: Restrict deprecated `DynamicConfig::rewrite()` usage
+- Enh #58: Restrict update by max core version of at least one installed module + Update modules
+
+2.3.0 (April 10, 2025)
+----------------------
+- Enh #55: For 1.18, Add a warning about the migration which will switch to the default theme and disable Theme Builder module
+
+2.2.2 (November 13, 2024)
 -------------------------
 - Fix #44: Hide a footer loader and an interrupt warning on process error
+- Enh #50: Use PHP CS Fixer
+- Enh #51: Reduce translation message categories
 
 2.2.1  (October 16, 2023)
 -------------------------